refactor(main.js): Simplify score updating and winner showing
Simplified score updating and put it in a function
(updateScore) in order to make the code more clear
and clean.
Simplified winner showing and put it in a function
(showWinner) in order to make the code more clear
and clean.
Also created a function (updateScoreAndShowWinner)
in order to call updateScore and showWinner
together